fix for recipe link
fixed migration files to not have recipeingredient table

for the migration to work you need to enableautomigration and
allowDataLoss.  After running update-database and running the SqlQuery_1
script I change it back and it is working for me.  So maybe only when
migrations are needed will we need to change these flags.

Added SQL query to create [IngredientRecipe] table with constraints to
other tables.

added ingredient collection to Recipe

Added linq query for many to many relationship for recipe and ingredient
in OnModelCreating